+++ changelog.md
@@ -2,13 +2,19 @@
 
 ---
 
-## [4.3.2] 2023-03-09
+## [4.4.0] 2023-03-09
 
 ### Changed
 
-- Update deps
+- Improved deployment of static assets prior to deploying app changes
+- Updated deps
 - Internal change: refactored globbing logic for Glob v9
 - Fixed typo; thanks @lpsinger!
+
+
+### Fixed
+
+- Fixed case where static assets would be pruned prior to deploying app changes, which could leave the app in a broken state during deployment, or if the deployment otherwise fails; fixes #1416
 
 ---